feat: expose account identifier utils
