+++ src/index.ts
@@ -9,4 +9,5 @@ export * from "./types/governance";
 export * from "./types/governance_converters";
 export * from "./types/icp";
 export * from "./types/ledger";
+export * from "./utils/account_identifier.utils";
 export * from "./utils/neurons.utils";
