+++ src/utils/converter.utils.ts
@@ -1,10 +1,7 @@
-import { Principal } from "@dfinity/principal";
 import { Buffer } from "buffer";
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 // @ts-ignore (no type definitions for crc are available)
 import crc from "crc";
-import { sha224 } from "js-sha256";
-import { AccountIdentifier } from "../types/common";
 
 export const uint8ArrayToBigInt = (array: Uint8Array): bigint => {
   const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
@@ -48,39 +45,6 @@ export const numberToArrayBuffer = (
 
 export const asciiStringToByteArray = (text: string): Array<number> => {
   return Array.from(text).map((c) => c.charCodeAt(0));
-};
-
-export const accountIdentifierToBytes = (
-  accountIdentifier: AccountIdentifier
-): Uint8Array => {
-  return Uint8Array.from(Buffer.from(accountIdentifier, "hex")).subarray(4);
-};
-
-export const accountIdentifierFromBytes = (
-  accountIdentifier: Uint8Array
-): AccountIdentifier => {
-  return Buffer.from(accountIdentifier).toString("hex");
-};
-
-export const principalToAccountIdentifier = (
-  principal: Principal,
-  subAccount?: Uint8Array
-): string => {
-  // Hash (sha224) the principal, the subAccount and some padding
-  const padding = asciiStringToByteArray("\x0Aaccount-id");
-
-  const shaObj = sha224.create();
-  shaObj.update([
-    ...padding,
-    ...principal.toUint8Array(),
-    ...(subAccount ?? Array(32).fill(0)),
-  ]);
-  const hash = new Uint8Array(shaObj.array());
-
-  // Prepend the checksum of the hash and convert to a hex string
-  const checksum = calculateCrc32(hash);
-  const bytes = new Uint8Array([...checksum, ...hash]);
-  return toHexString(bytes);
 };
 
 export const toHexString = (bytes: Uint8Array) =>
